Refactor JSON object representation
